create CheatableStarknetFacade
