+++ protostar/cairo_testing/execution_environments/cairo_setup_execution_environment.py
@@ -26,7 +26,7 @@ class CairoSetupExecutionEnvironment(CairoExecutionEnvironment):
     def _get_hint_locals(self, state: CairoTestExecutionState) -> HintLocalsDict:
         hint_locals: HintLocalsDict = {}
         cheatcode_factory = CairoSetupCheatcodeFactory(
-            cheatable_state=state.cheatable_state,
+            cheatable_state=state.starknet.cheatable_state,
             project_compiler=state.project_compiler,
         )
         cheatcodes = cheatcode_factory.build_cheatcodes()
